+++ nautilus_trader/adapters/interactive_brokers/client/common.py
@@ -534,7 +534,7 @@ class BaseMixin:
     # MarketData
     _bar_type_to_last_bar: dict[str, BarData | None]
     _bar_timeout_tasks: dict[str, Any]  # asyncio.Task
-    _order_id_to_order_ref: dict[int, AccountOrderRef]
+    _order_id_to_order_ref: dict[int | str, AccountOrderRef]
 
     # Order
     _next_valid_order_id: int